requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.2
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
GetOldTweets3==0.0.11
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lab==6.2
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pypiwin32==223
pyquery==1.4.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2020.4
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
tornado==6.1
tqdm==4.52.0
traitlets==5.0.5
translator==0.0.9
tweepy==3.9.0
txt2tags==3.7
tzlocal==2.1
urllib3==1.25.11
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
